+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Correctly compute spatialReference field
+
 ## [1.2.0] - 2017-02-10
 ### Added
 * Support for `outSR`